refactor: add scroll remember hook in preparation for use in other CardList pages
