+++ pages/index.tsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from 'react'
 import { NextPage, GetStaticProps } from 'next'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
 
 import { Service, getServices } from '../cms/services'
 import { CardList } from '../src/Components/CardList'
 import { HeaderComponent } from '../src/Components/Header'
 import { Layout } from '../src/Components/Layout'
 import { Welcome } from '../src/Components/Welcome'
+import { useScrollRemember } from '../src/hooks/scrollRemember'
 
 interface ListPageProps {
   services: Array<ServicePreview>
@@ -29,7 +29,7 @@ export type ServicePreview = Pick<
 export const ListPage: NextPage<ListPageProps> = ({ services }) => {
   // Default to showing the welcome screen
   const [showWelcome, setShowWelcome] = useState<boolean>(true)
-  const router = useRouter()
+  const { resumePosition } = useScrollRemember()
 
   useEffect(() => {
     // Attempt to retrieve the showWelcome value from local storage
@@ -38,22 +38,6 @@ export const ListPage: NextPage<ListPageProps> = ({ services }) => {
     // If it's been correctly set to false, update state
     stored === 'false' ? setShowWelcome(false) : setShowWelcome(true)
   }, [showWelcome])
-
-  // Store last scroll position so we can return to it when we load the page next time
-  useEffect(() => {
-    const handler = () => {
-      localStorage.setItem('lastScroll', window.pageYOffset.toString())
-    }
-
-    router.events.on('beforeHistoryChange', handler)
-    return () => {
-      router.events.off('beforeHistoryChange', handler)
-    }
-  }, [router.events])
-
-  const scrollToLastPos = () => {
-    window.scrollTo(0, parseInt(localStorage.getItem('lastScroll') || '0'))
-  }
 
   return (
     <>
@@ -68,7 +52,7 @@ export const ListPage: NextPage<ListPageProps> = ({ services }) => {
           <CardList
             services={services}
             listType="filtered"
-            onLoad={scrollToLastPos}
+            onLoad={resumePosition}
           />
         </Layout>
       )}
